waiting page files deleted

+++ app/src/main/java/com/example/optivote/model/UserDto.kt
@@ -3,7 +3,7 @@ package com.example.optivote.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UserDto (val id : Int? = null,
+data class UserDto (val id : Long? = null,
                val name:String? = null,
                val email:String? = null,
                val phone:String? = null,
