+++ app/src/main/java/com/example/optivote/model/decisionToSend.kt
@@ -3,6 +3,8 @@ package com.example.optivote.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class decisionToSend (val decision: String,
-                      val voteIdFk: Long,
-                      val userIdFk: Int)
+data class decisionToSend(
+    val decision: String,
+    val voteIdFk: Long,
+    val userIdFk: Long
+)
